refactor(mocks): little changing imports
